+++ app/src/main/java/com/example/beatlesapp/BeatlesApp.kt
@@ -1,4 +1,4 @@
-package com.example.beatlesapp.ui.theme
+package com.example.beatlesapp
 
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.Scaffold
@@ -9,6 +9,9 @@ import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
 import com.example.beatlesapp.data.AlbumList
+import com.example.beatlesapp.ui.screens.AlbumsScreen
+import com.example.beatlesapp.ui.screens.InfoScreen
+import com.example.beatlesapp.ui.screens.Routes
 
 @Composable
 fun BeatlesApp() {
