+++ app/src/main/java/com/example/beatlesapp/MainActivity.kt
@@ -7,7 +7,6 @@ import androidx.activity.enableEdgeToEdge
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.material3.Surface
 import androidx.compose.ui.Modifier
-import com.example.beatlesapp.ui.theme.BeatlesApp
 import com.example.beatlesapp.ui.theme.BeatlesAppTheme
 
 class MainActivity : ComponentActivity() {
