+++ app/src/main/java/com/example/beatlesapp/ui/screens/AlbumsScreen.kt
@@ -1,4 +1,4 @@
-package com.example.beatlesapp.ui.theme
+package com.example.beatlesapp.ui.screens
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
